Add CodeBuild badge integration w/ shields.io
The integration is based on Lambda (API and event processing),
EventBridge (CodeBuild build status events) and DynamoDB (persist build
state changes).

It's still a first-version code. It doesn't handle status by branche,
and it doesn't enforce stronger security controls.

Some comments regarding security: we don't check if the build project has
build badge enabled, and we don't enforce the requester matching the UUID
string (current API parameter is the build project name). UUID string is a
string CodeBuild already creates and requires in its native badge feature. Note
that requesting the UUID would allow us not to expose the underlying
CodeBuild project name and would allow us to only expose the build status if
build badge was enabled and if the requester knows the correct UUID).

This is why currently we're proposing to restrict the EventBridge rule
to only the build projects you want to expose. Check:
`extras/lambda/codebuild-badges-eventbridge/eventbridge-rule-template.json`.

But it's a functional code.