Add a new approach/solution for codebuild-badges extra
The new solution is based on proxying the native CodeBuild badge
feature.

Main goal is to re-use existing logic (built-in in CodeBuild) and
address security considerations explained in commit
1e592f1072f45d58df14f54cf621b98314c834af.

This new approach also comes with its own tradeoffs: it's more secure
since it has the same requirements (and the same parameters) of
CodeBuild badges, only showing enabled badges for valid uuids; but we
had to add uuid validation controls to avoid the API acting as an open
proxy - and because the validation relies on the CodeBuild control
plane APIs, we had to implement caching controls. So the solution has
some new complexities.